moved Store to its own module
